requirements.txt
matplotlib==3.9.2
numpy==2.1.1
pillow==10.4.0

